pi/requirements.txt
paho-mqtt>=1.6.0
pyserial>=3.5
